requirements.txt
# Django and REST framework versions
Django>=2.0.4,<5.0
djangorestframework>=3.8.2,<3.15
pytz>=2018.4

-e .

# Test requirements
pytest-django>=3.2.1,<=4.7.0
pytest>=3.5.1,<=7.4.4
pytest-cov
flake8>=3.5.0,<=7.0.0

# wheel for PyPI installs
wheel
twine>=1.11.0,<=4.0.2

# MkDocs for documentation previews/deploys
mkdocs
